requirements.txt
# Django & DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-storages==1.14.2
boto3==1.34.0

# Celery & Redis
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# PDF Processing
PyPDF2==3.0.1
pdfminer.six==20221105
pypdf==3.17.0

# Vector DB & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# LLM
openai==1.6.1
anthropic==0.18.1
transformers==4.36.2
torch==2.1.2

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
pydantic==2.5.3

# Development
pytest==7.4.3
pytest-django==4.7.0
black==23.12.1
flake8==6.1.0

# Jupyter
jupyter==1.0.0
ipykernel==6.26.0
notebook==7.0.6

